refactor: address PR review comments on txn events
