+++ Sources/Rokt_Widget/Networking/V2EventsClient.swift
@@ -3,7 +3,6 @@ import Foundation
 internal struct V2EventsClient {
     let baseURL: URL
     let accountId: String
-    let authToken: String?
     let sdkVersion: String
     let deviceHeaders: [String: String]
     let httpClient: HTTPClientAdapter
@@ -11,20 +10,18 @@ internal struct V2EventsClient {
     init(
         baseURL: URL,
         accountId: String,
-        authToken: String? = nil,
         sdkVersion: String,
         deviceHeaders: [String: String] = [:],
         httpClient: HTTPClientAdapter = RoktHTTPClient()
     ) {
         self.baseURL = baseURL
         self.accountId = accountId
-        self.authToken = authToken
         self.sdkVersion = sdkVersion
         self.deviceHeaders = deviceHeaders
         self.httpClient = httpClient
     }
 
-    func recordEvents(events: [V2Event]) async throws -> (Data?, HTTPURLResponse?) {
+    func recordEvents(events: [V2Event], authToken: String?) async throws -> (Data?, HTTPURLResponse?) {
         let url = baseURL
             .appendingPathComponent("v2")
             .appendingPathComponent("sessions")
