+++ Tests/ContractTests/V2EventsClientPactSpec.swift
@@ -4,7 +4,7 @@ import PactSwift
 
 /// Consumer-driven pact spec for the v2 `/v2/sessions/events` endpoint.
 ///
-/// Drives `V2EventsClient.recordEvents(events:)` — the test never constructs
+/// Drives `V2EventsClient.recordEvents(events:authToken:)` — the test never constructs
 /// request headers or body directly, only domain inputs. Wire-shape
 /// construction lives entirely in `V2EventsClient`, so any drift there
 /// (e.g. changing the `channel.type` body field) gets rejected by the pact mock service.
@@ -86,7 +86,6 @@ class V2EventsClientPactSpec: XCTestCase {
                     let client = V2EventsClient(
                         baseURL: url,
                         accountId: "account-456",
-                        authToken: "Bearer session-token-abc",
                         sdkVersion: "5.2.2"
                     )
                     let event = V2Event(
@@ -95,7 +94,10 @@ class V2EventsClientPactSpec: XCTestCase {
                         timestamp: 1_774_474_053_000,
                         data: ["source_message_id": "source-message-001"]
                     )
-                    let (_, httpResponse) = try await client.recordEvents(events: [event])
+                    let (_, httpResponse) = try await client.recordEvents(
+                        events: [event],
+                        authToken: "Bearer session-token-abc"
+                    )
                     XCTAssertEqual(httpResponse?.statusCode, 202)
                 } catch {
                     XCTFail("V2EventsClient request failed: \(error)")
